// src/lib/huawei/services/dds.types.ts
export type DdsInstance = {
  availabilityZone: string;
  backupKeepDays: string;
  backupWindow: string;
  chargeMode: string;
  createdAt: string;
  datastore: string;
  enterpriseProjectId: string;
  flavor: string;
  groups: Array<{
    id: string;
    name: string;
    nodes: Array<{
      availabilityZone: string;
      groupId: string;
      groupName: string;
      groupType: string;
      id: string;
      name: string;
      privateIp: string;
      role: string;
      status: string;
    }>;
    role: string;
    status: string;
    storage: string;
    storageType: string;
    type: string;
  }>;
  id: string;
  maintenanceWindow: string;
  mode: string;
  name: string;
  nodes: number;
  port: string;
  privateIp: string;
  publicIp: string;
  projectId: string;
  projectName: string;
  region: string;
  securityGroupId: string;
  status: string;
  storage: string;
  storageType: string;
  subnetId: string;
  switchStrategy: string;
  timeZone: string;
  type: string;
  updatedAt: string;
  vpcId: string;
};